+++ dos9/command/Dos9_CommandInfo.c
@@ -74,7 +74,8 @@ COMMANDINFO lpCmdInfo[]= {
 	{"WC", Dos9_CmdWc, STRLEN("WC")},
 	{"XARGS", Dos9_CmdXargs, STRLEN("XARGS")},
 	{"PROMPT", Dos9_CmdPrompt, STRLEN("PROMPT")},
-	{"PECHO", Dos9_CmdPecho, STRLEN("PECHO")}
+	{"PECHO", Dos9_CmdPecho, STRLEN("PECHO")},
+	{"TIMEOUT", Dos9_CmdTimeout, STRLEN("TIMEOUT")}
 };
 
 const int iCmdInfoNb = sizeof(lpCmdInfo)/sizeof(lpCmdInfo[0]);